implemented csrf-token

+++ App/Controllers/CategoriesController.php
@@ -25,7 +25,7 @@ class CategoriesController extends Controller {
     }
 
     public function add() {
-        if(!empty($_POST)) {
+        if(!empty($_POST) And Auth::checkCSRF($_POST["token"])) {
             $title       = isset($_POST['title']) ? $_POST['title'] : '';
             $description = isset($_POST['description']) ? $_POST['description'] : '';
 
@@ -57,6 +57,7 @@ class CategoriesController extends Controller {
                     ]
                 ]);
             }
+            App::redirect('categories');
         }
 
         else {
@@ -69,7 +70,7 @@ class CategoriesController extends Controller {
     }
 
     public function edit($id) {
-        if(!empty($_POST)) {
+        if(!empty($_POST) And Auth::checkCSRF($_POST["token"])) {
             $title       = isset($_POST['title']) ? $_POST['title'] : '';
             $description = isset($_POST['description']) ? $_POST['description'] : '';
 
@@ -133,7 +134,7 @@ class CategoriesController extends Controller {
         $model2 = new ProductsModel($_COOKIE['user']);
         $products = $model2->getProductsByCategoryId($id);
         
-        if(!empty($_POST)) {
+        if(!empty($_POST) And Auth::checkCSRF($_POST["token"])) {
             foreach($products as $product){
                 $model2->delete($product->id);
             }
